Fix ObjectHydrator on an edge case
In progress...

ObjectHydrator can fail if more than one row is returned from a JOINed result set and a root entity appears later on the row data. The problem is that the first time the root entity appears on the first row, it is marked on `$this->rootAliases`. On subsequent rows, the marked root entity is assumed to appear on the `$this->resultPointers` array which is not true if the entity data appears later on the raw row data.

TO-DO: Add failing and pass tests